Move masteries import into another file as it's slow as hell and sometimes we need to quickly re-run import.py
