+++ import.py
@@ -36,7 +36,6 @@ os.environ['DJANGO_SETTINGS_MODULE'] = 'evething.settings'
 from django.db import connections
 
 from thing.models import *  # NOPEP8
-from thing.helpers import roman
 
 # ---------------------------------------------------------------------------
 # Override volume for ships, assembled volume is mostly useless :ccp:
@@ -120,7 +119,6 @@ class Importer:
         time_func('NPCFaction', self.import_npcfaction)
         time_func('NPCCorporation', self.import_npccorporation)
         time_func('SkillMap', self.build_skill_map)
-        time_func('Masteries', self.import_masteries)
 
     # -----------------------------------------------------------------------
     # Regions
@@ -867,105 +865,6 @@ class Importer:
         cPickle.dump(skill_map, open('skill_map.pickle', 'w'))
 
         return 1
-
-    # -----------------------------------------------------------------------
-    def import_masteries(self):
-        # Masteries
-        added = 0
-
-        # Delete existing mastery skillplans
-        SkillPlan.objects.filter(visibility=SkillPlan.MASTERY_VISIBILITY).delete()
-
-        self.cursor.execute("""
-            SELECT s.skillID, s.skillLevel, m.masteryLevel, t.typeID, t.typeName
-            FROM certSkills AS s
-            JOIN certMasteries AS m ON s.certId = m.certID AND s.certLevelInt = m.masteryLevel
-            JOIN invTypes AS t on m.typeId = t.typeID
-        """)
-
-        ship_masteries = {}
-        for row in self.cursor:
-            if row[1] > 0:
-                if row[3] not in ship_masteries:
-                    ship_masteries[row[3]] = {
-                        'typeID': row[3],
-                        'name': row[4],
-                        'masteries': {}
-                    }
-                if row[2] not in ship_masteries[row[3]]['masteries']:
-                    ship_masteries[row[3]]['masteries'][row[2]] = {}
-
-                if row[0] in ship_masteries[row[3]]['masteries'][row[2]]:
-                    if row[1] > ship_masteries[row[3]]['masteries'][row[2]][row[0]]:
-                        ship_masteries[row[3]]['masteries'][row[2]][row[0]] = row[1]
-                else:
-                    ship_masteries[row[3]]['masteries'][row[2]][row[0]] = row[1]
-
-        for type_id in ship_masteries:
-            entries = []
-            for mastery_level in ship_masteries[type_id]['masteries']:
-                # print('\t%s - Mastery %s' % (ship_masteries[type_id]['name'], roman(mastery_level + 1)))
-
-                skillplan = SkillPlan.objects.create(
-                    name='%s - Mastery %s' % (ship_masteries[type_id]['name'], roman(mastery_level + 1)),
-                    visibility=SkillPlan.MASTERY_VISIBILITY
-                )
-
-                seen = {}
-                position = 0
-
-                for skill_id in ship_masteries[type_id]['masteries'][mastery_level]:
-                    level = ship_masteries[type_id]['masteries'][mastery_level][skill_id]
-
-                    # Get prereqs
-                    prereqs = Skill.get_prereqs(skill_id)
-                    for pre_skill_id, pre_level in prereqs:
-                        for i in range(seen.get(pre_skill_id, 0) + 1, pre_level + 1):
-                            # print('\t%d:\t%d' % (pre_skill_id, i))
-
-                            try:
-                                sps = SPSkill.objects.create(
-                                    skill_id=pre_skill_id,
-                                    level=i,
-                                    priority=3,
-                                )
-                            except:
-                                continue
-
-                            entries.append(SPEntry(
-                                skill_plan=skillplan,
-                                position=position,
-                                sp_skill=sps,
-                            ))
-
-                            position += 1
-                            seen[pre_skill_id] = i
-
-                    # Add the actual skill
-                    for i in range(seen.get(skill_id, 0) + 1, level + 1):
-                        # print('\t%d:\t%d' % (skill_id, i))
-                        try:
-                            sps = SPSkill.objects.create(
-                                skill_id=skill_id,
-                                level=i,
-                                priority=3,
-                            )
-                        except:
-                            continue
-
-                        entries.append(SPEntry(
-                            skill_plan=skillplan,
-                            position=position,
-                            sp_skill=sps,
-                        ))
-
-                        position += 1
-                        seen[skill_id] = i
-
-                added += 1
-            SPEntry.objects.bulk_create(entries)
-
-        return added
 
 # ---------------------------------------------------------------------------
